Create space.py
I know there is a cleaner way to write this code, but as of right now I don't know how. I'm still learning. When running the original print values for the weight it returned really long unrounded numbers on a few of the planets. I wanted to round it down to 2 decimal places so that it was easier to read. This code will show the original value, and the rounded value below for each planet.