Update space.py
Sometimes your weight on a given planet would return an unrounded number. I wanted to make things easier to read. I wrote this code here that shows the original weight and rounded weight. I'm 100% there is an easier and faster way to do this, but I don't know that method yet. I'm still learning. In the future I can return to this and modify it.